backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.0.8
pillow
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyasn1-modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
PyJWT==2.8.0
redis==5.0.3
zope.interface==6.2
requests==2.26.0
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
websockets==13.0.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-two-factor-auth
django-otp
psycopg2-binary==2.9.9
pyotp
qrcode
python-dotenv
web3==7.2.0